requirements.txt
numpy
torch
torchvision
faiss-cpu
nltk
tqdm
matplotlib
Pillow
pycocotools